logica de iniciar sesion añadida

+++ backend/app.js
@@ -1,14 +1,29 @@
+//////////////////////////////////////////////////////////////
+// Dependencias
+//////////////////////////////////////////////////////////////
 const func = require("./aplanar.js");
 
 const express = require("express");
 const app = express();
 const mongoose = require("mongoose");
-app.use(express.json());
 const cors = require("cors");
-const { FaTheRedYeti } = require("react-icons/fa");
-app.use(cors());
+const bcrypt = require("bcryptjs");
+const cp = require("cookie-parser");
+const jwt = require("jsonwebtoken");
+
+//////////////////////////////////////////////////////////////
+// Midleware de express
+//////////////////////////////////////////////////////////////
+
+app.use(express.json());
+app.use(cors({origin:"http://localhost:3000", credentials:true}));
 app.set("view engine", "ejs");
 app.use(express.urlencoded({ extended: false }));
+app.use(cp());
+
+//////////////////////////////////////////////////////////////
+// conexion con la base de datos
+//////////////////////////////////////////////////////////////
 
 const mongoUrl =
     "mongodb+srv://1910096:u2NnyYPZGiGi7bTO@prueba.stlmnyp.mongodb.net/";
@@ -22,13 +37,90 @@ mongoose
     })
     .catch((e) => console.log(e));
 
+//////////////////////////////////////////////////////////////
+// Modelos de la base de datos
+//////////////////////////////////////////////////////////////
+
 require("./RepuestoDetalles");
 require("./MotoDetalles");
+require("./UsuarioDetalles.js");
 
 const Repuesto = mongoose.model("Repuestos");
 const Moto = mongoose.model("Motos");
+const Usuario = mongoose.model("Usuarios");
+
+//////////////////////////////////////////////////////////////
+// Verificar si el usuario inicio sesion
+//////////////////////////////////////////////////////////////
+
+const estaAutorizado = (req) => {
+    console.log("estoy verificando si esta autorizado")
+    console.log(req.cookies);
+    if (!req.cookies.JWT) {
+        return false;
+    }
+    
+    const verificacionToken = jwt.verify(req.cookies.JWT, "EmpanadaDeJamonYQueso");
+    
+    if (!verificacionToken) {
+        return false;
+    }
+    
+    req.payload = jwt.decode(req.cookies.JWT, "EmpanadaDeJamonYQueso")
+    return true;
+}
+
+//////////////////////////////////////////////////////////////
+// Rutas de autenticacion
+//////////////////////////////////////////////////////////////
+
+app.post("/iniciar-sesion", async (req, res) => {
+    console.log("quiero iniciar sesion");
+    try {
+        const usuario = await Usuario.findOne({Usuario: req.body.Usuario});
+        if (!usuario) {
+            res.status(400).json({error: "El usuario no existe"}).send();
+            return;
+        }
+    
+        const comparacionContraseñas = await bcrypt.compare(req.body.Contraseña, usuario.Contraseña);
+        
+        if (!comparacionContraseñas) {
+            res.status(400).json({error:"La contraseña no es correcta"}).send();
+            return;  
+        }
+
+        const contenido = {Usuario: usuario.Usuario};
+        const token = jwt.sign(contenido,"EmpanadaDeJamonYQueso");
+    
+        res.cookie("JWT",token,{httpOnly:true, maxAge:24*60*60*1000}).send();
+
+    } catch {
+        res.status(400).json({error:"ha ocurrido un error al procesar su solicitud"});
+    }
+})
+
+app.get("/cerrar-sesion" , async (req, res) => {
+    res.clearCookie("JWT").json({response: "se ha cerrado la sesion"}).send();
+})
+
+app.get("/comprobar-sesion", async (req,res) => {
+    if (estaAutorizado(req)) {
+        res.status(200).send();
+    } else {
+        res.status(400).send();
+    }
+})
+
+//////////////////////////////////////////////////////////////
+// Rutas para trabajar con los productos de la base de datos
+//////////////////////////////////////////////////////////////
+
 app.post("/register", async (req, res) => {
     try {
+        if (!estaAutorizado(req)) 
+            throw "no estas autorizado";
+
         if (req.body["tipo"] === "Repuestos") {
             await Repuesto.create(req.body["form"]["Información Basica"]);
         } else if (req.body["tipo"] === "Motos") {
@@ -42,17 +134,18 @@ app.post("/register", async (req, res) => {
     }
 });
 
-app.listen(5000, () => {
-    console.log("Servidor Iniciado");
-});
-
 app.get("/obtenerProductos", async (req, res) => {
     try {
+        if (!estaAutorizado(req)) 
+            throw "no estas autorizado";
+
         let elementoProducto = [];
         if (req.query.tipo === "Repuestos") {
             elementoProducto = await Repuesto.find({});
         } else if (req.query.tipo === "Motos") {
             elementoProducto = await Moto.find({});
+        } else {
+            throw "no existe ese tipo de producto";
         }
         res.status(200).json(elementoProducto).send();
     } catch (error) {
@@ -62,6 +155,9 @@ app.get("/obtenerProductos", async (req, res) => {
 
 app.post("/eliminarProducto", async (req, res) => {
     try {
+        if (!estaAutorizado(req)) 
+            throw "no estas autorizado";
+
         if (req.body["tipo"] === "Repuestos") {
             Repuesto.deleteOne(
                 { "Código de parte": req.body["datos"]["Código de parte"] },
@@ -96,6 +192,9 @@ app.post("/eliminarProducto", async (req, res) => {
 
 app.post("/suspensionProducto", async (req, res) => {
     try {
+        if (!estaAutorizado(req)) 
+            throw "no estas autorizado";
+
         if (req.body["tipo"] === "Repuestos") {
             Repuesto.findOneAndUpdate(
                 { "Código de parte": req.body["datos"]["Código de parte"] },
@@ -134,6 +233,9 @@ app.post("/agregacionMasivaProducto", async (req, res) => {
     let sesion = await mongoose.startSession();
     sesion.startTransaction();
     try {
+        if (!estaAutorizado(req)) 
+            throw "no estas autorizado";
+
         if (req.body["tipo"] === "Repuestos") {
             await Repuesto.insertMany(req.body["datos"], { sesion });
         } else if (req.body["tipo"] === "Motos") {
@@ -149,4 +251,12 @@ app.post("/agregacionMasivaProducto", async (req, res) => {
     } finally {
         await sesion.endSession();
     }
+});
+
+//////////////////////////////////////////////////////////////
+// Puerto del servidor
+//////////////////////////////////////////////////////////////
+
+app.listen(5000, () => {
+    console.log("Servidor Iniciado");
 });
